POC / hack to fix non-function-templates.cpp test
as a POC, implement hack to create a new FunctionDecl that
substitutes template parameters into function paramter types, but no
into the return type. This allows downstream logic in
findInstantiationOf to find the ParmVarDecl mapping. A potential proper
fix for this is to create a version of SubstDecl that only substitutes
into the parameters of the function type. Another option is to leverage
SubstParmTypes, and insert the mapping that findInstantiationOf expects
(Not sure how to do this).