[tsan] Introduce Adaptive Delay Scheduling to TSAN
This commit introduces an "adaptive delay scheduler" to the TSAN
runtime. At various synchronization points, like mutex or atomic
operations, the scheduler may sleep or do small spin loops to expose
additional thread interleavings in the execution.

This change is inspired by prior work, which is discussed in more detail
on https://discourse.llvm.org/t/rfc-tsan-implementing-a-fuzz-scheduler-for-tsan/80969.
In short, https://reviews.llvm.org/D65383 was an earlier unmerged
attempt at adding a random delay delay scheduler. Feedback on the RFC
led to the version in this commit, aiming to limit the amount of delay.

This change aims to limit the overhead when the fuzzing scheduler is not
enabled (empty TSAN_OPTIONS). Nearly all checks are guarded by an
inlined checked on a global variable that is only set when the fuzzing
scheduler is active. I ran some contrived benchmarks with empty
TSAN_OPTIONS with an unmodified TSAN runtime and with the TSAN runtime
with these changes. For example, a simple main() that creates one
background thread and runs millions of atomic ops (relaxed, acq/rel,
seq_cst) in a tight loop), and did not find meaningful (or unmeaningful)
differences in the performance.

The scheduler is added as a virtual interface to allow other behaviors
discussed in https://ceur-ws.org/Vol-2344/paper9.pdf to be added in the
future (for example, a scheduler that only allows one thread to run at
once).

An LLM helped write the adaptive behavior, in particular, the TimeBudget
class for program-level delay tracking, the tiering concept for how much
to delay the program based on type of synchronization operation, the
address sampler, and the per-thread quota. I reviewed the ouput and
amended the code to reduce some duplication and simplify some of the
behavior. I also used the LLM to replace its original `double`
based calculation logic with the Percent class, allowing integer based
arithmetic. Finally, the LLM helped write the unit test cases for
Percent.