Attempting to integrate with BlockingCall
Instead of a custom mutex/cv implementation, the idea was to have
BlockingCall mark the current thread as BLOCKED, and mark the thread
RUNNING when the blocking call finished. However, one problem that
occurs is:

 - Thread A owns a mutex in RUNNING state. At a sync point,
 - Thread B is chosen which then attempts to lock the mutex. It marks
   itself BLOCKED, then asks thread A to run
 - Thread A continues, then waits on a condition. Before the
   the thread enters the real pthread_cond_wait, it marks itself
   as BLOCKED and tries to mark another thread as runnable. Since
   thread B is still BLOCKED, no threads can run.

We might be able to work around this by tracking which threads are
blocked on a given lock, and which threads are about to wait on a
cv associated with the same mutex, but this could get complicated,
and is a subset of the work required to support the bespoke mutex/cv
implementation anyway.